requirements.txt
python-docx==1.1.2
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.1.1
requests==2.32.3
werkzeug==3.0.1

openai==1.102.0
groq==0.31.1
zhipuai>=2.1.5
google-generativeai==0.7.2
dashscope==1.24.2
volcengine-python-sdk[ark]==4.0.14
volcengine==1.0.166

pandas==2.2.3
numpy==2.2.1
scipy==1.15.0
matplotlib==3.10.0
seaborn==0.13.2
datasets==4.0.0

PyPDF2==3.0.1
pymupdf==1.23.26
markitdown[all]==0.1.3
pyaudio==0.2.14
nbformat==5.10.4

pydantic==2.11.7
pydantic-settings==2.5.2
python-json-logger==2.0.7
prometheus-client==0.20.0
psutil==5.9.8
aiofiles==23.2.0
tqdm==4.67.1
nltk==3.9.1
arxiv==2.2.0
Kaleido==1.1.0
redis==5.0.3
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-mock==3.14.0
black==24.4.2
flake8==7.1.0
mypy==1.10.0
nest_asyncio==1.5.6
plotly==6.3.0
tabulate==0.9.0
fastmcp==2.12.2
scholarly==1.7.11
crawl4ai==0.7.4


